improvement(DEV-2628): put for sale balance check
