+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@refinableco/refinable-sdk",
-  "version": "5.0.0-next.6",
+  "version": "5.0.0-next.7",
   "description": "The Refinable SDK allows you to easily interact with the Refinable Marketplace to mint, sell and buy NFTs",
   "main": "./lib/index.js",
   "types": "./lib/index.d.ts",
@@ -22,7 +22,7 @@
     "assert": "^2.0.0",
     "axios": "^0.25.0",
     "eth-rpc-errors": "^4.0.3",
-    "ethers": "^5.7.0",
+    "ethers": "^5.7.2",
     "eventemitter3": "^4.0.7",
     "form-data": "^4.0.0",
     "graphql": "^16.6.0",
